requirements.txt
pyyaml==6.0.1
requests
openapi3